+++ apps/blog/next-env.d.ts
@@ -1,3 +1,3 @@
 /// <reference types="next" />
 /// <reference types="next/types/global" />
-/// <reference types="styled-components/cssprop" />
+/// <reference types="next/image-types/global" />